[#github_4199] Change TCPReceiver to interface
To extend the usability of the TCPReceiver, change it to interface.
