+++ collector/src/main/java/com/navercorp/pinpoint/collector/receiver/tcp/TCPReceiver.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 NAVER Corp.
+ * Copyright 2018 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,80 +16,13 @@
 
 package com.navercorp.pinpoint.collector.receiver.tcp;
 
-import com.navercorp.pinpoint.collector.receiver.PinpointServerAcceptorProvider;
-import com.navercorp.pinpoint.rpc.server.PinpointServerAcceptor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.net.InetSocketAddress;
-import java.util.Objects;
-import java.util.concurrent.Executor;
-
 /**
  * @author Taejin Koo
  */
-public class TCPReceiver {
+public interface TCPReceiver {
 
-    private final Logger logger;
+    void start();
 
-    private final String name;
-
-    private final InetSocketAddress bindAddress;
-    private final PinpointServerAcceptorProvider acceptorProvider;
-
-    private PinpointServerAcceptor serverAcceptor;
-
-    private final Executor executor;
-
-    private final TCPPacketHandler tcpPacketHandler;
-
-
-    public TCPReceiver(String name, TCPPacketHandler tcpPacketHandler, Executor executor, InetSocketAddress bindAddress, PinpointServerAcceptorProvider acceptorProvider) {
-        this.name = Objects.requireNonNull(name, "name must not be null");
-        this.logger = LoggerFactory.getLogger(name);
-
-        this.bindAddress = Objects.requireNonNull(bindAddress, "bindAddress must not be null");
-
-        this.acceptorProvider = Objects.requireNonNull(acceptorProvider, "acceptorProvider must not be null");
-        this.executor = Objects.requireNonNull(executor, "executor must not be null");
-
-        this.tcpPacketHandler = Objects.requireNonNull(tcpPacketHandler, "tcpPacketHandler must not be null");
-
-    }
-
-    public void start() {
-        if (logger.isInfoEnabled()) {
-            logger.info("{} start() started", name);
-        }
-        final PinpointServerAcceptor acceptor = newAcceptor();
-        acceptor.bind(bindAddress);
-        this.serverAcceptor = acceptor;
-        if (logger.isInfoEnabled()) {
-            logger.info("{} start() completed", name);
-        }
-    }
-
-    private PinpointServerAcceptor newAcceptor() {
-        PinpointServerAcceptor acceptor = acceptorProvider.get();
-
-        // take care when attaching message handlers as events are generated from the IO thread.
-        // pass them to a separate queue and handle them in a different thread.
-        acceptor.setMessageListenerFactory(new TCPReceiverServerMessageListenerFactory(executor, tcpPacketHandler));
-        return acceptor;
-    }
-
-    public void shutdown() {
-        if (logger.isInfoEnabled()) {
-            logger.info("{} shutdown() started", name);
-        }
-
-        if (serverAcceptor != null) {
-            serverAcceptor.close();
-        }
-
-        if (logger.isInfoEnabled()) {
-            logger.info("{} shutdown() completed", name);
-        }
-    }
+    void shutdown();
 
 }
